fix #3 crash over 2147MHz
Thanks to "Jake dev" --> https://github.com/LamaBleu/plutoscripts/issues/3

+++ pluto_power/pow.c
@@ -286,7 +286,7 @@ int main (int argc, char **argv)
             rxgain = (int16_t) atoi(optarg);
             break;
         case 'l':
-            lo_mhz = (long long) atoi(optarg);
+            lo_mhz = (long long) atol(optarg);
             break;
         case 'f':
             fs_mhz = (long) atoi(optarg);
